requirements.txt
asgiref==3.5.0
Django==4.0.3
django-model-utils==4.0.0
djangorestframework==3.13.1
psycopg2-binary
pytz==2021.3
sqlparse==0.4.2
django-simple-history==3.0.0